pyproject.toml
[project]
name = "matrx-microservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.12",
    "python-socketio==5.13.0",
    "websockets==12.0",
    "uvicorn==0.31.0",
    "supabase>=2.15.0",
    "pydantic-settings>=2.9.1",
    "concurrent-log-handler>=0.9.25",
]
